+++ src/extension.ts
@@ -1,95 +1,7 @@
 import * as vscode from "vscode";
-import { Language, Parser } from "web-tree-sitter";
-
-const LINES_BEFORE = 2;
-const LINES_AFTER = 2;
-
-type ReferenceGroup = { uri: vscode.Uri; locations: vscode.Location[] };
-
-function groupByUri(locations: vscode.Location[]): ReferenceGroup[] {
-  const results: { uri: vscode.Uri; locations: vscode.Location[] }[] = [];
-  let uri: vscode.Uri | undefined;
-  let locs: vscode.Location[] = [];
-  for (const loc of locations) {
-    if (!uri || uri.toString() !== loc.uri.toString()) {
-      if (uri && locs.length > 0) {
-        results.push({ uri, locations: locs });
-      }
-      uri = loc.uri;
-      locs = [];
-    }
-    locs.push(loc);
-  }
-  if (uri && locs.length > 0) {
-    results.push({ uri, locations: locs });
-  }
-  return results;
-}
-
-class StreamingFile {
-  stream: AsyncGenerator<string, void, unknown>;
-  onUpdate: () => void;
-  private readonly chunks: string[] = [];
-  constructor(
-    stream: AsyncGenerator<string, void, unknown>,
-    onUpdate: () => void,
-  ) {
-    this.stream = stream;
-    this.onUpdate = onUpdate;
-  }
-
-  async startStream() {
-    for await (const chunk of this.stream) {
-      this.chunks.push(chunk);
-      this.onUpdate();
-    }
-  }
-
-  stopStream() {
-    this.stream.return();
-  }
-
-  get content() {
-    return this.chunks.join("\n");
-  }
-}
-
-class VirtualFileProvider
-  implements vscode.TextDocumentContentProvider, vscode.Disposable {
-  scheme: string;
-  private contentMap: Map<string, StreamingFile> = new Map();
-  private index = 0;
-  private emitter = new vscode.EventEmitter<vscode.Uri>();
-
-  readonly onDidChange = this.emitter.event;
-
-  constructor(scheme: string) {
-    this.scheme = scheme;
-  }
-
-  addContent(contentStream: AsyncGenerator<string>): vscode.Uri {
-    const uri = vscode.Uri.parse(`${this.scheme}:${this.index++}.code-search`);
-    const streamingFile = new StreamingFile(contentStream, () =>
-      this.emitter.fire(uri),
-    );
-    this.contentMap.set(uri.toString(), streamingFile);
-    streamingFile.startStream();
-    return uri;
-  }
-  removeContent(uri: vscode.Uri) {
-    this.contentMap.get(uri.toString())?.stopStream();
-    this.contentMap.delete(uri.toString());
-  }
-  provideTextDocumentContent(uri: vscode.Uri): vscode.ProviderResult<string> {
-    return this.contentMap.get(uri.toString())?.content;
-  }
-
-  dispose() {
-    for (const streamingFile of this.contentMap.values()) {
-      streamingFile.stopStream();
-    }
-  }
-}
+import {Language, Parser} from "web-tree-sitter";
+import {VirtualFileProvider} from "./virtualFileProvider.js";
+import {renderReferences} from "./references.js";
 
 async function loadParser(context: vscode.ExtensionContext) {
   await Parser.init(
@@ -129,7 +41,7 @@ export function activate(context: vscode.ExtensionContext) {
         "xrefs.findAllXrefs",
         async (editor) => {
           const uri = fileProvider.addContent(
-            renderResults(
+            renderReferences(
               vscode.commands.executeCommand<vscode.Location[]>(
                 "vscode.executeReferenceProvider",
                 editor.document.uri,
@@ -147,89 +59,3 @@ export function activate(context: vscode.ExtensionContext) {
   })();
 }
 
-async function* renderResults(
-  references: Thenable<vscode.Location[]>,
-): AsyncGenerator<string, void, unknown> {
-  const refs = await references;
-
-  const refGroups = groupByUri(refs);
-
-  yield `Found ${refs.length} xrefs in ${refGroups.length} files.\n`;
-
-  for (const refGroup of refGroups) {
-    yield renderReferenceGroup(refGroup);
-  }
-}
-async function renderReferenceGroup(refGroup: ReferenceGroup): Promise<string> {
-  const doc = await vscode.workspace.openTextDocument(refGroup.uri);
-  const lines = [];
-  const refs = refGroup.locations.toSorted((a, b) =>
-    a.range.start.compareTo(b.range.start),
-  );
-  const path = refGroup.uri.fsPath;
-
-  lines.push(`${path}:`);
-  type LineType = "context" | "reference" | "spacer";
-  /*
-  For some odd reason, `new Map()` does not work in the remote extension host.
-  So instead of maps, we're using objects.
-  It's terrible, but it works.
-  */
-  const resultLines: Record<
-    number,
-    { type: LineType; line: number; text: string }
-  > = Object.create(null);
-  let maxLine = 0;
-  for (const ref of refs.slice(1)) {
-    const space_before = Math.max(0, ref.range.start.line - LINES_BEFORE - 1);
-    resultLines[space_before] = {
-      type: "spacer",
-      line: space_before,
-      text: "",
-    };
-  }
-  for (const ref of refs) {
-    const before = Math.max(0, ref.range.start.line - LINES_BEFORE);
-    const after = Math.min(doc.lineCount, ref.range.end.line + LINES_AFTER + 1);
-    for (let line = before; line < after; ++line) {
-      console.log("Line", line);
-      resultLines[line] = {
-        type: "context",
-        line,
-        text: doc.lineAt(line).text,
-      };
-    }
-
-    maxLine = Math.max(maxLine, after);
-  }
-  for (const ref of refs) {
-    const line = ref.range.start.line;
-    console.log("Line:", line);
-    resultLines[line] = {
-      type: "reference",
-      line,
-      text: doc.lineAt(line).text,
-    };
-  }
-
-  const padCount = `${maxLine + 1}`.length;
-  const pad = (n: number) => `${n}`.padStart(padCount);
-  for (const res of Object.values(resultLines).toSorted(
-    (a, b) => a.line - b.line,
-  )) {
-    switch (res.type) {
-      case "spacer":
-        lines.push("");
-        break;
-      case "context":
-        lines.push(`  ${pad(res.line + 1)}  ${res.text}`);
-        break;
-      case "reference":
-        lines.push(`  ${pad(res.line + 1)}: ${res.text}`);
-    }
-  }
-  lines.push("");
-
-  const content = lines.join("\n");
-  return content;
-}
